H210068T MICHAEL KARIMBIKA/backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Faker==36.1.1
idna==3.10
joblib==1.4.2
jsonpickle==4.0.2
numpy==1.26.4
pandas==2.2.3
pesepay==1.0.8
pillow==11.1.0
pycryptodome==3.22.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scikit-surprise==1.1.4
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.1
urllib3==2.3.0
